# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Human Dashboard API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/play/show-yardage/assistant-coach": {
            "post": {
                "tags": [
                    "Broadcast"
                ],
                "summary": "Head coach → assistant: broadcast play run so AC opens the position-after-play modal",
                "operationId": "yardagePlayToAssistant",
                "requestBody": {
                    "required": false,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/YardagePlayToAssistantRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Broadcast dispatched to assistant coach",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/YardagePlayToAssistantResponse"
                                },
                                "example": {
                                    "status": 200,
                                    "message": "Yardage play broadcast sent to assistant coach.",
                                    "data": {
                                        "head_coach_id": 5,
                                        "channel": "coach-group.5.league.22",
                                        "event": "assistant.coaches",
                                        "payload": {
                                            "playName": "IN 25(4)-RAM ALL-CHECK",
                                            "yardageGain": 45,
                                            "sliderDirection": "ltr",
                                            "targetTeam": "offensive",
                                            "suggestionData": {
                                                "yardage": 10,
                                                "positionNumber": 45,
                                                "quarter": 1,
                                                "down": 2,
                                                "target": "12",
                                                "myteamId": 1,
                                                "oppteamId": 2,
                                                "strategies": "regular",
                                                "weather_status": "Normal",
                                                "game_id": 36,
                                                "league_id": 22,
                                                "time": "12:00",
                                                "my_points": 7,
                                                "oponent_points": 0,
                                                "mode": "practice"
                                            },
                                            "selectedPlayIds": 280,
                                            "play": {
                                                "id": 280,
                                                "title": "IN 25(4)-RAM ALL-CHECK"
                                            },
                                            "type": "offensive",
                                            "targetPlayers": [
                                                {
                                                    "value": 101,
                                                    "text": "RB Player"
                                                }
                                            ],
                                            "my_team": 1,
                                            "opponent_team": 2,
                                            "mode": "practice"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Caller not linked to a head coach group",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Head coach group is not available for this user."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Pusher broadcast failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Broadcast failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/assistant-coach/system-suggestion/broadcast": {
            "post": {
                "tags": [
                    "Broadcast"
                ],
                "summary": "Assistant coach → head coach: broadcast scoreboard context for System Suggestions",
                "operationId": "assistantSystemSuggestionBroadcast",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/AssistantSystemSuggestionBroadcastRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Broadcast dispatched to head coach",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/AssistantSystemSuggestionBroadcastResponse"
                                },
                                "example": {
                                    "status": 200,
                                    "message": "System suggestion broadcast sent to head coach.",
                                    "data": {
                                        "head_coach_id": 5,
                                        "channel": "coach-group.5.league.22",
                                        "event": "head.coach.suggestion",
                                        "payload": {
                                            "down": 2,
                                            "weather": "Rain",
                                            "strategies": "regular",
                                            "expected_yardage_gain": 10,
                                            "h_mark_position": "hmark_center",
                                            "game_id": 36,
                                            "league_id": 22,
                                            "mode": "practice",
                                            "actor_id": 15,
                                            "actor_name": "Assistant Coach"
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden — caller is not an assistant coach",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Forbidden"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error or assistant not linked to a head coach",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "Head coach is not linked to this assistant."
                                                }
                                            },
                                            "type": "object"
                                        },
                                        {
                                            "properties": {
                                                "message": {
                                                    "type": "string",
                                                    "example": "The given data was invalid."
                                                },
                                                "errors": {
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Pusher broadcast failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Broadcast failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/configurations/{key}": {
            "get": {
                "tags": [
                    "Configuration"
                ],
                "summary": "Get a single configuration by key",
                "description": "Returns one configuration entry for the given key.",
                "operationId": "getConfigurationByKey",
                "parameters": [
                    {
                        "name": "key",
                        "in": "path",
                        "description": "Configuration key",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "RPP_CONFIGURATION"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Configuration entry",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "key": {
                                            "type": "string",
                                            "pattern": "^[A-Za-z0-9_]+$",
                                            "example": "RPP_CONFIGURATION"
                                        },
                                        "value": {
                                            "type": "string",
                                            "maxLength": 255,
                                            "pattern": "^[A-Za-z0-9_]+$",
                                            "example": "with_rpp"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Configuration not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/leagues/{league}/qb/logout": {
            "post": {
                "tags": [
                    "League QB"
                ],
                "summary": "Head coach force-logout league QB",
                "operationId": "logoutLeagueQb",
                "parameters": [
                    {
                        "name": "league",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 22
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer",
                                        "example": 78
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "QB logged out"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "YardagePlayToAssistantResponse": {
                "properties": {
                    "status": {
                        "type": "integer",
                        "example": 200
                    },
                    "message": {
                        "type": "string",
                        "example": "Yardage play broadcast sent to assistant coach."
                    },
                    "data": {
                        "properties": {
                            "head_coach_id": {
                                "type": "integer",
                                "example": 5
                            },
                            "league_id": {
                                "type": "integer",
                                "example": 22
                            },
                            "channel": {
                                "type": "string",
                                "example": "coach-group.5.league.22"
                            },
                            "event": {
                                "type": "string",
                                "example": "assistant.coaches"
                            },
                            "payload": {
                                "$ref": "#/components/schemas/YardagePlayBroadcastPayload"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Sanctum bearer token. Use format: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}